% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrichment_analysis.R
\name{run_tissue_enrichment}
\alias{run_tissue_enrichment}
\title{Run Tissue Enrichment Analysis}
\usage{
run_tissue_enrichment(
  input_genes,
  background_genes = NULL,
  specificity_types = c("Tissue enriched", "Group enriched", "Tissue enhanced")
)
}
\arguments{
\item{input_genes}{Vector of gene symbols.}

\item{background_genes}{Vector of background gene symbols.}

\item{specificity_types}{Character vector of HPA categories to use.}
}
\description{
Run Tissue Enrichment Analysis
}
